fixup! Добавила иконку для гугл-авторизации

+++ UseBy/Screens/MainAuth/MainAuthView.swift
@@ -17,13 +17,26 @@ class MainAuthView: UIView {
     private let appDescriptionLabel = UILabel()
     private let googleSignUpButton: MainButton
     private let signUpButton = MainButton(text: "create-account".localized, theme: MainButton.ButtonTheme.action)
-    private let alreadySignUpButton = MainButton(text: "already-have-account".localized, theme: MainButton.ButtonTheme.clear)
-    private let backgroundCircle = BackgroundCircle(frame: CGRect(x: CGFloat(0), y: CGFloat(0), width: CGFloat(500), height: CGFloat(500)), circleColor: UIColor.mainActionBGColor())
-    private let bottomBackgroundCircle = BackgroundCircle(frame: CGRect(x: CGFloat(0), y: CGFloat(0), width: CGFloat(600), height: CGFloat(600)), circleColor: UIColor.secondaryActionBGColor())
+    private let alreadySignUpButton = MainButton(
+        text: "already-have-account".localized,
+        theme: MainButton.ButtonTheme.clear
+    )
+    private let backgroundCircle = BackgroundCircle(
+        frame: CGRect(x: 0, y: 0, width: 500, height: 500),
+        circleColor: UIColor.mainActionBGColor()
+    )
+    private let bottomBackgroundCircle = BackgroundCircle(
+        frame: CGRect(x: 0, y: 0, width: 600, height: 600),
+        circleColor: UIColor.secondaryActionBGColor()
+    )
 
     init() {
         let googleIcon = Icon.getIcon(name: "Google", size: .small, color: UIColor.inversedTextColor())
-        googleSignUpButton = MainButton(text: "create-account-google".localized, theme: MainButton.ButtonTheme.social, icon: googleIcon)
+        googleSignUpButton = MainButton(
+            text: "create-account-google".localized,
+            theme: MainButton.ButtonTheme.social,
+            icon: googleIcon
+        )
         super.init(frame: .zero)
 
         backgroundColor = UIColor.mainBGColor()
